* trying to verify 64-bit validity of code, checking casts, especially when pointers are involved * first pass of converting c-style casts to c++ casts (static|const|reinterpret|dynamic) where possible * using misc.h to_string and from_string more often * removed unnecessary casts * using size_t more often, instead of casting to and from int/unsigned int/unsigned long * upped clientsmlcsharp project to warning level 4, fixed warnings

+++ KernelSML/src/sml_InputListener.cpp
@@ -42,7 +42,7 @@ void InputListener::OnKernelEvent(int eventID, AgentSML* pAgentSML, void* pCallD
 	switch (eventID) {
 		case smlEVENT_INPUT_PHASE_CALLBACK:
 		{
-			int callbacktype = (int)reinterpret_cast<long long>(pCallData);
+			int callbacktype = reinterpret_cast<int>(pCallData);
 
 			switch(callbacktype) {
 			case TOP_STATE_JUST_CREATED:
